Refactor HappyStoriesList and HappyStoriesListItems
- All files named in CamelCase
- optimised code
