+++ right-for-life_front/src/app/HappyStoriesList/HappyStoriesList.js
@@ -1,14 +1,10 @@
 import React, { useState, useEffect } from 'react';
-import { HappyStoriesListItem } from '../happy-stories-list-item';
+import { HappyStoriesListItem } from '../HappyStoriesListItem';
 
 export function HappyStoriesList(props) {
 
   const [ happyStoriesData, setHappyStoriesData ] = useState(null);
   const [ isStoriesDisplayed, setIsStoriesDisplayed ] = useState(false);
-
-  function handleShowStoriesButtonClick() {
-    setIsStoriesDisplayed(!isStoriesDisplayed);
-  }
 
   useEffect(() => {
     fetch('https://raw.githubusercontent.com/AlexeyKasaev3/softServe-academy/master/demo-3-data/news.json')
@@ -18,16 +14,16 @@ export function HappyStoriesList(props) {
 
   return <div>
     <button
-      onClick={handleShowStoriesButtonClick}
+      onClick={() => setIsStoriesDisplayed(!isStoriesDisplayed)}
     >
       { isStoriesDisplayed ? 'Hide Stories' : 'Show Stories' }
     </button>
     {isStoriesDisplayed && happyStoriesData && (
       <>
         <h1>Истории</h1>
-        {happyStoriesData.map(story => {
-          return <HappyStoriesListItem image={story.image} title={story.title} />
-        })}
+        {happyStoriesData.map( story =>
+          <HappyStoriesListItem key={story.id} image={story.image} title={story.title} />
+        )}
       </>
     )
     }
