+++ right-for-life_front/src/app/HappyStoriesList/index.js
@@ -1,0 +1,1 @@
+export { HappyStoriesList } from "./HappyStoriesList";
