+++ right-for-life_front/src/app/HappyStoriesListItem/index.js
@@ -1,0 +1,1 @@
+export { HappyStoriesListItem } from "./HappyStoriesListItem";
