--- right-for-life_front/src/app/happy-stories-list-item/happyStoriesListItem.js
@@ -1,8 +1,0 @@
-import React from 'react'
-
-export function HappyStoriesListItem(props) {
-  return <article>
-    <img src={props.image} alt={props.title}/>
-    <h2><a href="somewhere">{props.title}</a></h2>
-  </article>
-}
