--- right-for-life_front/src/app/happy-stories-list-item/index.js
@@ -1,1 +1,0 @@
-export { HappyStoriesListItem } from "./happyStoriesListItem";
