--- right-for-life_front/src/app/happy-stories-list/index.js
@@ -1,1 +1,0 @@
-export { HappyStoriesList } from "./happyStoriesList";
